cleanup

+++ lib/data/sources/repositories/compose_repository_impl.dart
@@ -31,10 +31,12 @@ class ComposeRepositoryImpl extends ComposeRepository {
       [bool? allowUnconfirmedTx, int? fee]) async {
     final response = await api.composeSendVerbose(
         sourceAddress, destination, asset, quantity, allowUnconfirmedTx, fee);
+
     if (response.result == null) {
       // TODO: handle errors
       throw Exception('Failed to compose send');
     }
+
     final txVerbose = response.result!;
     return compose_send.ComposeSend(
         params: compose_send.ComposeSendParams(
@@ -115,7 +117,7 @@ class ComposeRepositoryImpl extends ComposeRepository {
         params: compose_issuance.ComposeIssuanceVerboseParams(
           source: txVerbose.params.source,
           asset: txVerbose.params.asset,
-          quantity: txVerbose.params.quantity,
+          quantity: txVerbose.params.quantity.toDouble(),
           divisible: txVerbose.params.divisible,
           lock: txVerbose.params.lock,
           description: description,
